Admin: renames Chart.min.js as chart.3.8.0.min.js
Linux users reported that the plugin failed to load chart.min.js on
their browsers. This was caused due to the filename being Chart.min.js
on the server while the plugin attempts to load chart.min.js

Renamed the file as chart.3.8.0.min.js to fix the issue and also so I
can easily keep track of the version that's being used with the plugin.